greeting files changed to navbar
