+++ app/views/layouts/application.html.erb
@@ -4,8 +4,6 @@
   <title>ScareBnB</title>
   <%= stylesheet_link_tag    'application', media: 'all' %>
   <%= javascript_include_tag 'application' %>
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
-
   <%= csrf_meta_tags %>
 </head>
 <body>
